# src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Sample Ecommerce App
  description: Ecommerce API developed designed using OpenAPI specifications
tags:
  - name: customer
    description: operations about customers
  - name: hello
    description: operations about saying hello   	
    
paths:
  
  '/api/v1/hello':
    get:
      tags: 
        - hello
      summary: returns hello
      description: says hello world
      operationId: sayHello
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/HelloMessage'
            application/json:
              schema:
                $ref: '#/components/schemas/HelloMessage'

  '/api/v1/customers':
    get:
      tags: 
        - customer
      summary: returns all customers
      description: returns all customers or empty collection if not found
      operationId: getAllCustomers
      responses:
        200:
          description: for successfull fetch
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
                  
  '/api/v1/customers/{id}':
    get:
      tags:
        - customer
      summary: returns a customer
      description: returns a customer identifiable by a given ID
      operationId: getCustomerById
      parameters:
       - name: id
         in: path
         description: customer identifier
         required: true
         schema:
           type: integer
           format: int64
      responses:
        200:
          description: For successful fetch
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Customer'
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

components:
  schemas:
    
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
      xml:
        name: Customer
    
    HelloMessage:
      type: object
      properties:
        headerStr:
          type: string
        contentStr:
          type: string
      xml:
        name: HelloMessage      
                                     
                
                